refactored laplegend files to have filenames match purpose; adjusting layout as well
